refactor: select
