+++ components/vc-select/SelectTrigger.tsx
@@ -1,19 +1,12 @@
 import Trigger from '../vc-trigger';
 import PropTypes from '../_util/vue-types';
-import { getSlot } from '../_util/props-util';
 import classNames from '../_util/classNames';
-import createRef from '../_util/createRef';
 import type { CSSProperties } from 'vue';
-import { defineComponent } from 'vue';
-import type { RenderDOMFunc } from './interface';
-import type { DropdownRender } from './interface/generator';
-import type { Placement } from './generate';
+import { computed, ref, defineComponent } from 'vue';
 import type { VueNode } from '../_util/type';
+import type { DropdownRender, Placement, RenderDOMFunc } from './BaseSelect';
 
-const getBuiltInPlacements = (dropdownMatchSelectWidth: number | boolean) => {
-  // Enable horizontal overflow auto-adjustment when a custom dropdown width is provided
-  const adjustX = typeof dropdownMatchSelectWidth !== 'number' ? 0 : 1;
-
+const getBuiltInPlacements = (adjustX: number) => {
   return {
     bottomLeft: {
       points: ['tl', 'bl'],
@@ -49,6 +42,19 @@ const getBuiltInPlacements = (dropdownMatchSelectWidth: number | boolean) => {
     },
   };
 };
+
+const getAdjustX = (
+  adjustXDependencies: Pick<SelectTriggerProps, 'autoAdjustOverflow' | 'dropdownMatchSelectWidth'>,
+) => {
+  const { autoAdjustOverflow, dropdownMatchSelectWidth } = adjustXDependencies;
+  if (!!autoAdjustOverflow) return 1;
+  // Enable horizontal overflow auto-adjustment when a custom dropdown width is provided
+  return typeof dropdownMatchSelectWidth !== 'number' ? 0 : 1;
+};
+export interface RefTriggerProps {
+  getPopupElement: () => HTMLDivElement;
+}
+
 export interface SelectTriggerProps {
   prefixCls: string;
   disabled: boolean;
@@ -66,103 +72,122 @@ export interface SelectTriggerProps {
   getPopupContainer?: RenderDOMFunc;
   dropdownAlign: object;
   empty: boolean;
+  autoAdjustOverflow?: boolean;
   getTriggerDOMNode: () => any;
+  onPopupVisibleChange?: (visible: boolean) => void;
+
+  onPopupMouseEnter: () => void;
 }
 
 const SelectTrigger = defineComponent<SelectTriggerProps, { popupRef: any }>({
   name: 'SelectTrigger',
   inheritAttrs: false,
-  created() {
-    this.popupRef = createRef();
-  },
+  props: {
+    dropdownAlign: PropTypes.object,
+    visible: PropTypes.looseBool,
+    disabled: PropTypes.looseBool,
+    dropdownClassName: PropTypes.string,
+    dropdownStyle: PropTypes.object,
+    placement: PropTypes.string,
+    empty: PropTypes.looseBool,
+    autoAdjustOverflow: PropTypes.looseBool,
+    prefixCls: PropTypes.string,
+    popupClassName: PropTypes.string,
+    animation: PropTypes.string,
+    transitionName: PropTypes.string,
+    getPopupContainer: PropTypes.func,
+    dropdownRender: PropTypes.func,
+    containerWidth: PropTypes.number,
+    dropdownMatchSelectWidth: PropTypes.oneOfType([Number, Boolean]).def(true),
+    popupElement: PropTypes.any,
+    direction: PropTypes.string,
+    getTriggerDOMNode: PropTypes.func,
+    onPopupVisibleChange: PropTypes.func,
+    onPopupMouseEnter: PropTypes.func,
+  } as any,
+  setup(props, { slots, attrs, expose }) {
+    const builtInPlacements = computed(() => {
+      const { autoAdjustOverflow, dropdownMatchSelectWidth } = props;
+      return getBuiltInPlacements(
+        getAdjustX({
+          autoAdjustOverflow,
+          dropdownMatchSelectWidth,
+        }),
+      );
+    });
+    const popupRef = ref();
+    expose({
+      getPopupElement: () => {
+        return popupRef.value;
+      },
+    });
+    return () => {
+      const { empty = false, ...restProps } = { ...props, ...attrs };
+      const {
+        visible,
+        dropdownAlign,
+        prefixCls,
+        popupElement,
+        dropdownClassName,
+        dropdownStyle,
+        direction = 'ltr',
+        placement,
+        dropdownMatchSelectWidth,
+        containerWidth,
+        dropdownRender,
+        animation,
+        transitionName,
+        getPopupContainer,
+        getTriggerDOMNode,
+        onPopupVisibleChange,
+        onPopupMouseEnter,
+      } = restProps as SelectTriggerProps;
+      const dropdownPrefixCls = `${prefixCls}-dropdown`;
 
-  methods: {
-    getPopupElement() {
-      return this.popupRef.current;
-    },
-  },
+      let popupNode = popupElement;
+      if (dropdownRender) {
+        popupNode = dropdownRender({ menuNode: popupElement, props });
+      }
 
-  render() {
-    const { empty = false, ...props } = { ...this.$props, ...this.$attrs };
-    const {
-      visible,
-      dropdownAlign,
-      prefixCls,
-      popupElement,
-      dropdownClassName,
-      dropdownStyle,
-      direction = 'ltr',
-      placement,
-      dropdownMatchSelectWidth,
-      containerWidth,
-      dropdownRender,
-      animation,
-      transitionName,
-      getPopupContainer,
-      getTriggerDOMNode,
-    } = props as SelectTriggerProps;
-    const dropdownPrefixCls = `${prefixCls}-dropdown`;
+      const mergedTransitionName = animation ? `${dropdownPrefixCls}-${animation}` : transitionName;
 
-    let popupNode = popupElement;
-    if (dropdownRender) {
-      popupNode = dropdownRender({ menuNode: popupElement, props });
-    }
+      const popupStyle = { minWidth: `${containerWidth}px`, ...dropdownStyle };
 
-    const builtInPlacements = getBuiltInPlacements(dropdownMatchSelectWidth);
-
-    const mergedTransitionName = animation ? `${dropdownPrefixCls}-${animation}` : transitionName;
-
-    const popupStyle = { minWidth: `${containerWidth}px`, ...dropdownStyle };
-
-    if (typeof dropdownMatchSelectWidth === 'number') {
-      popupStyle.width = `${dropdownMatchSelectWidth}px`;
-    } else if (dropdownMatchSelectWidth) {
-      popupStyle.width = `${containerWidth}px`;
-    }
-    return (
-      <Trigger
-        {...props}
-        showAction={[]}
-        hideAction={[]}
-        popupPlacement={placement || (direction === 'rtl' ? 'bottomRight' : 'bottomLeft')}
-        builtinPlacements={builtInPlacements}
-        prefixCls={dropdownPrefixCls}
-        popupTransitionName={mergedTransitionName}
-        popup={<div ref={this.popupRef}>{popupNode}</div>}
-        popupAlign={dropdownAlign}
-        popupVisible={visible}
-        getPopupContainer={getPopupContainer}
-        popupClassName={classNames(dropdownClassName, {
-          [`${dropdownPrefixCls}-empty`]: empty,
-        })}
-        popupStyle={popupStyle}
-        getTriggerDOMNode={getTriggerDOMNode}
-      >
-        {getSlot(this)[0]}
-      </Trigger>
-    );
+      if (typeof dropdownMatchSelectWidth === 'number') {
+        popupStyle.width = `${dropdownMatchSelectWidth}px`;
+      } else if (dropdownMatchSelectWidth) {
+        popupStyle.width = `${containerWidth}px`;
+      }
+      return (
+        <Trigger
+          {...props}
+          showAction={[]}
+          hideAction={[]}
+          popupPlacement={placement || (direction === 'rtl' ? 'bottomRight' : 'bottomLeft')}
+          builtinPlacements={builtInPlacements.value}
+          prefixCls={dropdownPrefixCls}
+          popupTransitionName={mergedTransitionName}
+          popupAlign={dropdownAlign}
+          popupVisible={visible}
+          getPopupContainer={getPopupContainer}
+          popupClassName={classNames(dropdownClassName, {
+            [`${dropdownPrefixCls}-empty`]: empty,
+          })}
+          popupStyle={popupStyle}
+          getTriggerDOMNode={getTriggerDOMNode}
+          onPopupVisibleChange={onPopupVisibleChange}
+          v-slots={{
+            default: slots.default,
+            popup: () => (
+              <div ref={popupRef} onMouseenter={onPopupMouseEnter}>
+                {popupNode}
+              </div>
+            ),
+          }}
+        ></Trigger>
+      );
+    };
   },
 });
-
-SelectTrigger.props = {
-  dropdownAlign: PropTypes.object,
-  visible: PropTypes.looseBool,
-  disabled: PropTypes.looseBool,
-  dropdownClassName: PropTypes.string,
-  dropdownStyle: PropTypes.object,
-  placement: PropTypes.string,
-  empty: PropTypes.looseBool,
-  prefixCls: PropTypes.string,
-  popupClassName: PropTypes.string,
-  animation: PropTypes.string,
-  transitionName: PropTypes.string,
-  getPopupContainer: PropTypes.func,
-  dropdownRender: PropTypes.func,
-  containerWidth: PropTypes.number,
-  dropdownMatchSelectWidth: PropTypes.oneOfType([Number, Boolean]).def(true),
-  popupElement: PropTypes.any,
-  direction: PropTypes.string,
-  getTriggerDOMNode: PropTypes.func,
-};
 
 export default SelectTrigger;
