+++ components/vc-select/Selector/Input.tsx
@@ -16,7 +16,7 @@ interface InputProps {
   autofocus: boolean;
   autocomplete: string;
   editable: boolean;
-  accessibilityIndex: number;
+  activeDescendantId?: string;
   value: string;
   open: boolean;
   tabindex: number | string;
@@ -45,7 +45,7 @@ const Input = defineComponent({
     autofocus: PropTypes.looseBool,
     autocomplete: PropTypes.string,
     editable: PropTypes.looseBool,
-    accessibilityIndex: PropTypes.number,
+    activeDescendantId: PropTypes.string,
     value: PropTypes.string,
     open: PropTypes.looseBool,
     tabindex: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
@@ -86,7 +86,7 @@ const Input = defineComponent({
       autofocus,
       autocomplete,
       editable,
-      accessibilityIndex,
+      activeDescendantId,
       value,
       onKeydown,
       onMousedown,
@@ -131,7 +131,7 @@ const Input = defineComponent({
           'aria-owns': `${id}_list`,
           'aria-autocomplete': 'list',
           'aria-controls': `${id}_list`,
-          'aria-activedescendant': `${id}_list_${accessibilityIndex}`,
+          'aria-activedescendant': activeDescendantId,
           ...attrs,
           value: editable ? value : '',
           readonly: !editable,
