+++ components/vc-select/TransBtn.tsx
@@ -1,10 +1,11 @@
 import type { FunctionalComponent } from 'vue';
 import type { VueNode } from '../_util/type';
 import PropTypes from '../_util/vue-types';
+import type { RenderNode } from './BaseSelect';
 
 export interface TransBtnProps {
   class: string;
-  customizeIcon: VueNode | ((props?: any) => VueNode);
+  customizeIcon: RenderNode;
   customizeIconProps?: any;
   onMousedown?: (payload: MouseEvent) => void;
   onClick?: (payload: MouseEvent) => void;
