+++ components/vc-select/interface.ts
