--- components/vc-select2/OptGroup.tsx
@@ -1,15 +1,0 @@
-import type { FunctionalComponent } from 'vue';
-
-import type { DefaultOptionType } from './Select';
-
-export type OptGroupProps = Omit<DefaultOptionType, 'options'>;
-
-export interface OptionGroupFC extends FunctionalComponent<OptGroupProps> {
-  /** Legacy for check if is a Option Group */
-  isSelectOptGroup: boolean;
-}
-
-const OptGroup: OptionGroupFC = () => null;
-OptGroup.isSelectOptGroup = true;
-OptGroup.displayName = 'ASelectOptGroup';
-export default OptGroup;
