--- components/vc-select2/utils/platformUtil.ts
@@ -1,4 +1,0 @@
-/* istanbul ignore file */
-export function isPlatformMac(): boolean {
-  return /(mac\sos|macintosh)/i.test(navigator.appVersion);
-}
