+++ components/vc-tree-select/index.tsx
@@ -1,3 +1,4 @@
+// base rc-tree-select@4.6.1
 import TreeSelect from './TreeSelect';
 import TreeNode from './TreeNode';
 import { SHOW_ALL, SHOW_CHILD, SHOW_PARENT } from './utils/strategyUtil';
